// src/lib/api/models/createMyNeedDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { CreateMyNeedDtoPurpose } from './createMyNeedDtoPurpose';
import type { CreateMyNeedDtoDynamicValuesJson } from './createMyNeedDtoDynamicValuesJson';

export interface CreateMyNeedDto {
  purpose: CreateMyNeedDtoPurpose;
  propertyTypeId?: string;
  provinceId?: string;
  districtId?: string;
  wardId?: string;
  budgetMin?: number;
  budgetMax?: number;
  areaMin?: number;
  areaMax?: number;
  bedrooms?: number;
  expectedTime?: string;
  note?: string;
  dynamicValuesJson?: CreateMyNeedDtoDynamicValuesJson;
}